kmip_host: Add gnutls prio string to config
Fixes #797

On centos 7, the very old gnutls (3.3) and/or the default prio settings
tries to use the "wrong" cert negotiation, and does not work with our
cryptsoft test host. While obviously not universal, this shows that
depending on OS and kmip target server, the user might need to
specialize the prio strings for kmip.

Message-Id: <20190617121849.16606-1-calle@scylladb.com>