workload prioritization: make workload prioritization internal queries run with timeout
This commit makes the workload prioritization run with
a timeout identical to the other tables in the distributed
system keyspace. This serves the purpos of preventing
from calls to the system_dystributed functions that are
related to workload prioritization to
to get stuck forever (or a very long time) under some
circumstances.
If the calls are stuck it can affect workload prioritization
configuration changes detection and appliance.