kmip_host: Enable keepalive and retry to deal with stale/lost connections
Fixes #998

If we spend time not using a connection, or the server simply goes down,
we can end up with a stale connection, which in turn will give us an
IO error (unlike a protocol error).

Add tcp keepalive to make connections last longer, and also retry
for commands on IO errors. This will also handle failover if one
server in group goes down.

Message-Id: <20190807164945.17588-1-calle@scylladb.com>