Remove Alternator from enterprise
NOTE: This is a patch to Scylla Enterprise main branch ("enterprise"),
not to the "alternator" branch!

Preparing to be released to open source, Alternator was moved to a
separate branch, "alternator". The version in the main enterprise
branch is already outdated, and may confuse unsuspecting testers to
test the wrong version. So this patch removes this outdated code.

Eventually as alternator reaches open source, the newest version of it
will be merged back into the main enterprise branch branch.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190811100400.31685-1-nyh@scylladb.com>