kmip_host: Do find of existing keys on server before creating new one
Fixes #1039

When asked for a key with ALG+strength (AES-128 etc), the KMIP provider always creates
a new key unless it is in the local attribute cache.

The cache is rather short-lived, and more to the point will not be populated until ID
lookup or by the create itself. I.e. on start we always create new keys for new
sstables.

To reduce the amount of keys generated on the KMIP server, we should do a server search
for matching (no yet invalidated) keys first, and reuse existing ID:s if possible.

This patch adds a lookup call before actual create.

Message-Id: <20190910111235.27229-1-calle@scylladb.com>