in-memory: Remove double command line registration
The in-memory-storage-in-mb command line have been
registered twice. Once in the named value constructor
and the other in the add_options function of the
in memory mirorred file class that created the named value.
This caused scylla to fail while starting whenever this
option was used.
This patch removes the registration from the add_options
function since registering in the named value itself is the
common way of registering.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20191225124841.20930-1-eliransin@scylladb.com>