encryption: Do not unconditionally assume "id" == uuid in sstable describe
Fixes #1211
Backport to 2019.*

When adding the "key_id" attribute, the code collecting sstable metadata assumed it
could be expressed as a UUID. However, for replicated provider the id is actually
a 33 byte vector, a length that makes UUID_gen assert and crash.

Changed to encode as UUID if possible, otherwise just hex string.

Message-Id: <20200115103620.2452-1-calle@scylladb.com>