cql3: grammar: drop raw cql text (for audit) from batchStatementObjective
Commit 7dc7ef0 ("audit: add query string to audit_info") added the raw
cql string to the statement. This causes antlr to generate a
batchStatementObjective_returns proxy object to hold the return
type. This proxy object has a copy constructor.

Commit adc6a98b5 ("cql3: return raw::parsed_statement as unique_ptr")
changed the return type of batchStatementObjective to be a unique_ptr.
This makes the bactchStatementObjective_return copy constructor fail
to compile.

To make forward progress, I am undoing the part of the first commit
that modifies the grammar, so the raw text will be empty, creating
a regression.

Opens #1307.