Workload Prioritization: return the correct error code for invalid request
The workload prioritization specific exception for invalid
requests which is usually thrown when trying to act on a none
existing service level, inherited from a wrong exception class.
The result of this was incorrect error code returned from Scylla,
the returned error code (0000) means internal server error, while
the actual error code should be 2200.
The bug was discovered in workload prioritization dtests, probabely
starting from a specific version of Cassandra driver which is more
strict about the error codes.
This commit fixes the inheritance which in turn causes Scylla to
return the correct error code for the invalid requests.

Tests: Unit tests - dev
       Service Level dtests (a modified version that expects the right
       error code)