build: stop using underscore in version number
Debian doesn't like underscores in version numbers [1]. We have hacks
to convert the underscore to something else, but we can just as easily
use a period instead. This gives us the same version numbers across
all packaging formats, and fewer hacks.

[1] https://www.debian.org/doc/debian-policy/ch-controlfields.html#version

I'll post patches to remove the hacks once this goes in here and in the
submodules.