features: add WORKLOAD_PRIORITIZATION feature
This commit add the WORKLOAD_PRIORITIZATION feature,
it is an addition that in itself is not very useful since
WP is already supported in previous versions, however,
this is added to prevent future complications related
to the fact that WP creates tables and might do so in a
mixed cluster. A follow up commit will be needed in order
to really support correct behaviour from previous versions
to this one.