distributed keyspace: don't cteate workload prioritization tables if cluster doesn't support them
Up until now, we created the workload prioritization tables
unconditionally. This could cause some issues when the
cluster is mixed and doesn't support workload prioritization.
This commit splits the system distributed keyspace initialization
routines in a way that will allow to start the workload prioritization
part later if the cluster currently doesn't support it.
The decision of currently support it is divided into two:
 1. in the context of the distributed keyspaces global start() function ,
 we will create the tables only if the WORKLOAD_PRIORITIZATION feature
 flag is enabled and the configuration doesn't disables the creation.
 2. in the context of initializing later (for example after the whole
 cluster has been upgraded) it is left to the user of the system distributed
 class and allow him to only start the workload prioritization part on demand.

The workload prioritization start routine can be called multiple times
and this will not affect anything if workload prioritization have
already been started.