workload prioritization: Only start workload prioritization when appropriate
Part of the workload prioritization mechanism activity is to create
tables and then periodically poll them for configuration changes.
This doesn't work well in a mixed cluster. The just created table
has different digests throughout the cluster, this causes errors
on read which are unnecessary.
This commit delays the table creation and polling until
when it is appropriate.
The appropriate time is when the whole cluster supports workload
prioritization, and this is indicated by either the WORKLOAD_PRIORITIZATION
feature flag or the pre existence of the workload prioritization tables.
The former should not be used and it is a workaround for the fact
that previous versions that supported the workload prioritization
feature didn't have the feature flag. We should eventually remove
this condition when the oldest supported version is the version
that contains this commit.

Tests:
  Tested the 3 different cases - that result in different code paths:
  1. Upgrading from a version without workload prioritization.
  2. Upgrading from a version with workload prioritization but
     without the feature flag.
  3. Installing a new cluster with this version.