Workload Prioritization: don't start polling the table before tokens initialization.
The workload prioritization polling loop gets "hooked" into
the table before the node has initialized it's token mapping.
The reason why this doesn't happen every time is that the workload
prioritization polling loop is set to 10 seconds, most of the time
the first polling occurs after the tokens have been initialized
but if the polling happens to be before, we get the above error as
a result of a scan query.

Fixes #1461