test.py: Wait for slapd to exit
After slapd receives a SIGTERM, it removes its pid file from the
instance directory before exiting.  But this happens concurrently with
test.py invoking rmtree to remove the whole instance directory.  There
is a race condition there; one possible consequence is a
FileNotFoundError in rmtree if slapd.pid is removed after rmtree globs
it but before it attempts to delete it.

To prevent the race condition, add a Popen.wait() call on the slapd
process, which should delay rmtree until after slapd has exited,
ensuring that slapd.pid is deleted.

Also, don't ignore errors when deleting the slapd instance directory.
It's important to know if this removal fails, as the directory
location interferes with the next test.py run.

Fixes #1466.

Tests: unit (dev)

Signed-off-by: Dejan Mircevski <dejan@scylladb.com>
Message-Id: <20200826211729.32260-1-dejan@scylladb.com>