sstables/ICS: Fix severe slowness of streaming in ICS during bootstrap
In the commit 87dd23db0311e960f0964, we fixed slowness of streaming during bootstrap
for all compaction strategies but ICS. This problem happens because the bootstrap
process can generate lots of relatively small SSTables that would cause a huge
write amplification in compaction, slowing down streaming. This problem is fixed
by temporarily increasing the min threshold, and reverting it back to the original
value once streaming completes. This will be better handled with off-strategy
compaction for streaming, but until we have that, let's provide ICS with this fix.

Fixes #1480.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20200902191035.20855-1-raphaelsc@scylladb.com>