encryption at rest: Create keyspace and table through migration manager
a544ca64e2 made internal queries not to be allowed to change the schema,
however, modules should and can create keyspaces and tables. The way to do it
is through the migration manager, we directly build the schema and then
announce it. The current implementation of encryption at rest replicated
key is trying to create the keyspace and table using internal cql queries
which are not not going to work.
This commit does not change the semantics but only use different
mechanisms to achieve the same goal of creating keyspace and tables for
replicated key provider of encryption at rest.

Fixes #1588

Closes #1584